requirements.txt
ddt==1.1.1
Jinja2==2.9.6
MarkupSafe==1.0
nose==1.3.7
nose-html-reporting==0.2.3
nosehtmloutput-2==1.0
selenium==3.6.0
xlrd==1.1.0